requirements/main.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile ./main.in -v
#
django==2.2
djangorestframework==3.9.2
gunicorn==19.9.0
python-decouple==3.1
pytz==2019.1              # via django
semantic-version==2.6.0
sqlparse==0.3.0           # via django
